lasīt izmaiņas README failā

+++ learning_platform/src/main/java/lv/slugboot/app/controller/StudentHomeController.java
@@ -14,6 +14,9 @@ import lombok.RequiredArgsConstructor;
 import lv.slugboot.app.models.Course;
 import lv.slugboot.app.models.LabInstance;
 import lv.slugboot.app.models.Student;
+import lv.slugboot.app.models.enums.LabInstanceStatus;
+import lv.slugboot.app.service.ICourseCRUDService;
+import lv.slugboot.app.service.ILabInstanceCRUDService;
 import lv.slugboot.app.service.IStudentHomeService;
 
 @Controller
@@ -22,6 +25,8 @@ import lv.slugboot.app.service.IStudentHomeService;
 public class StudentHomeController {
 
 	private final IStudentHomeService studentHomeService;
+	private final ICourseCRUDService courseCRUDService;
+	private final ILabInstanceCRUDService instanceCRUDService;
 
 	private static final String STUDENT_HOME_PAGE = "student-home-page";
 	private static final String ERROR_PAGE = "show-error";
@@ -61,12 +66,18 @@ public class StudentHomeController {
 		try {
 			String courseIdStr = request.getParameter(COURSE_ID_PARAMETER);
 			UUID courseId = UUID.fromString(courseIdStr);
-
+			Course course = courseCRUDService.retrieveById(courseId);
 			String username = authentication.getName();
 			Student student = studentHomeService.getStudentByUsername(username);
 			UUID studentId = student.getPersonId();
+			LabInstance instance = instanceCRUDService.retrieveByCourseAndStudent(course, student);
+			
 			model.addAttribute(STUDENT_ATTRIBUTE, student);
-
+			
+			if (instance.getStatus() != LabInstanceStatus.UNINITIALIZED) {
+				throw new IllegalArgumentException("Cannot remove a course until its instance is uninitialized");
+			}
+			
 			studentHomeService.removeCourseFromStudent(studentId, courseId);
 
 			Collection<Course> filteredCourses = studentHomeService.getAllCourses(studentId);
